non-chunking version
This version works better on RTX4060 with 8GB RAM and 48GB system memory. The lack of chunking on the system with smaller system memory enables 12s operation. The chunking version 5 requires 2x system memory of the array (at least 80GB) and will crash on a system with only 48GB. But if the array is too large for thrust, it will run out of memory and crash.